Move validation to a separate package
